requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
pytest
pydantic[email]
python-jose
python-jose[cryptography]
fastapi-security
fastapi-utils
bcrypt
redis
prometheus-fastapi-instrumentator